+++ src/core/admin.py
@@ -1,5 +1,3 @@
 from django.contrib import admin
 
-from core.models import CustomUser, Profile
 
-admin.site.register([CustomUser, Profile])
