+++ src/core/models.py
@@ -1,80 +1,3 @@
-from django.contrib.auth import get_user_model
-from django.contrib.auth.base_user import AbstractBaseUser
-from django.contrib.auth.models import PermissionsMixin
-from django.core.validators import FileExtensionValidator
-from django.db import models
-from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
-from phonenumber_field.modelfields import PhoneNumberField
-
-from core.managers import CustomerManager
-from core.utils.general import directory_path
 
 
-class CustomUser(AbstractBaseUser, PermissionsMixin):
-    first_name = models.CharField(_("name"), max_length=150, blank=True)
-    last_name = models.CharField(_("surname"), max_length=150, blank=True)
-    email = models.EmailField(_("email address"), unique=True)
-    is_staff = models.BooleanField(
-        _("staff status"),
-        default=False,
-        help_text=_("Designates whether the user can log into this admin site."),
-    )
-    is_active = models.BooleanField(
-        _("active"),
-        default=True,
-        help_text=_(
-            "Designates whether this user should be treated as active. " "Unselect this instead of deleting accounts."
-        ),
-    )
-    date_joined = models.DateTimeField(_("date joined"), default=timezone.now)
 
-    objects = CustomerManager()
-
-    USERNAME_FIELD = "email"
-    REQUIRED_FIELDS = []
-
-    class Meta:
-        verbose_name = _("user")
-        verbose_name_plural = _("users")
-
-    def clean(self):
-        super().clean()
-        self.email = self.__class__.objects.normalize_email(self.email)
-
-    def get_full_name(self):
-        """
-        Return the first_name plus the last_name, with a space in between.
-        """
-        full_name = f"{self.first_name} {self.last_name}"
-        return full_name.strip()
-
-    def get_short_name(self):
-        """Return the short name for the user."""
-        return self.first_name
-
-    def __str__(self):
-        return f"{self.first_name} {self.last_name} {self.email}"
-
-
-class Profile(models.Model):
-    user = models.OneToOneField(get_user_model(), on_delete=models.CASCADE)
-    avatar = models.ImageField(
-        upload_to=directory_path,
-        validators=[FileExtensionValidator(allowed_extensions=["jpg", "jpeg", "png"])],
-        null=True,
-        blank=True,
-        default="users/no_avatar.jpeg",
-    )
-    birth_date = models.DateField(null=True, blank=True)
-    phone_number = PhoneNumberField(_("phone number"), null=True, blank=True)
-
-    @property
-    def get_photo_url(self):
-        if self.avatar and hasattr(self.avatar, "url"):
-            return self.avatar.url
-        else:
-            return "users/no_avatar.jpeg"
-
-    def __str__(self):
-        return f"{self.user.pk} {self.user.first_name} {self.user.last_name} {self.user.email}"
