+++ src/core/services/signals.py
@@ -1,4 +1,4 @@
-from core.models import Profile
+from accounts.models import Profile
 
 
 def create_user_profile_signal(sender, instance, created, **kwargs):
